requirements.txt
Django==1.6.5
Unipath==1.0
django-crispy-forms==1.4.0
